+++ packages/react/src/components/UIShell/__stories__/UIShell.stories.js
@@ -7,13 +7,13 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React from 'react';
+import React, { useState } from 'react';
 import mdx from './UIShell.mdx';
-import { SideNav } from '../components/SideNav';
+import { SIDE_NAV_TYPE, SideNav } from '../components/SideNav';
 import { SideNavItems } from '../components/SideNavItems';
 import { SideNavMenu } from '../components/SideNavMenu';
 import { SideNavMenuItem } from '../components/SideNavMenuItem';
-
+import { HeaderPanel } from '../components/HeaderPanel';
 import {
   SideNavLink,
   SkipToContent,
@@ -24,9 +24,21 @@ import {
   HeaderMenuButton,
   SideNavDivider,
   Content,
+  HeaderGlobalBar,
+  HeaderGlobalAction,
 } from '@carbon/react';
-import { Fade, Add } from '@carbon/icons-react';
+import { Add, Fade, SquareOutline } from '@carbon/icons-react';
 import '../components/ui-shell.scss';
+
+export default {
+  title: 'Components/UIShell',
+  component: SideNav,
+  parameters: {
+    docs: {
+      page: mdx,
+    },
+  },
+};
 
 /**
  * Story content
@@ -77,6 +89,7 @@ const StoryContent = () => {
         used per product section. If tabs are needed on a page when using a
         side-nav, then the tabs are secondary in hierarchy to the side-nav.
       </p>
+      <HeaderPanel />
     </div>
   );
   const style = {
@@ -89,21 +102,130 @@ const StoryContent = () => {
   );
 };
 
-export default {
-  title: 'Components/UIShell',
-  component: SideNav,
-  parameters: {
-    docs: {
-      page: mdx,
-    },
-  },
+/**
+ * Story for UIShell
+ * @returns {React.ReactElement} The JSX for the story
+ */
+
+export const Default = () => {
+  const [expandedPanel, setExpandedPanel] = useState(false);
+  return (
+    <div>
+      <Theme theme="g100">
+        <HeaderContainer
+          render={({ isSideNavExpanded, onClickSideNavExpand }) => (
+            <>
+              <Header aria-label="IBM Platform Name">
+                <SkipToContent />
+                <HeaderMenuButton
+                  aria-label={isSideNavExpanded ? 'Close menu' : 'Open menu'}
+                  onClick={onClickSideNavExpand}
+                  isActive={isSideNavExpanded}
+                  aria-expanded={isSideNavExpanded}
+                  isCollapsible //shows hamburger menu at desktop
+                  isFixedNav
+                />
+                <HeaderName href="#" prefix="IBM">
+                  [Platform]
+                </HeaderName>
+                <HeaderGlobalBar>
+                  <HeaderGlobalAction
+                    aria-label={expandedPanel ? 'Close panel' : 'Open panel'}
+                    isActive={expandedPanel}
+                    aria-expanded={expandedPanel}
+                    tooltipAlignment="end"
+                    onClick={() => setExpandedPanel(!expandedPanel)}>
+                    <Fade size={20} />
+                  </HeaderGlobalAction>
+                </HeaderGlobalBar>
+                <HeaderPanel expanded={expandedPanel} />
+              </Header>
+              <SideNav
+                aria-label="Side navigation1"
+                expanded={isSideNavExpanded}
+                onSideNavBlur={onClickSideNavExpand}
+                isCollapsible
+                hideOverlay
+                className="nav--global">
+                <SideNavItems>
+                  <SideNavMenu renderIcon={SquareOutline} title="Link">
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                  </SideNavMenu>
+                  <SideNavMenu renderIcon={SquareOutline} title="Link">
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                  </SideNavMenu>
+                  <SideNavMenu renderIcon={SquareOutline} title="Link">
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                  </SideNavMenu>
+                  <SideNavMenu renderIcon={SquareOutline} title="Link">
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                  </SideNavMenu>
+                  <SideNavMenu renderIcon={SquareOutline} title="Link">
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                  </SideNavMenu>
+                  <SideNavDivider />
+                  <SideNavLink renderIcon={SquareOutline} href="#">
+                    Link
+                  </SideNavLink>
+                  <SideNavLink renderIcon={SquareOutline} href="#">
+                    Link
+                  </SideNavLink>
+                  <SideNavMenu renderIcon={SquareOutline} title="Link">
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                    <SideNavMenuItem href="#">Link</SideNavMenuItem>
+                  </SideNavMenu>
+                </SideNavItems>
+              </SideNav>
+              <SideNav
+                navType={SIDE_NAV_TYPE.PANEL}
+                isChildOfHeader={false}
+                aria-label="Side navigation">
+                <SideNavItems>
+                  <SideNavLink renderIcon={Fade} href="#">
+                    Link
+                  </SideNavLink>
+                  <SideNavLink renderIcon={Fade} href="#">
+                    Link
+                  </SideNavLink>
+                  <SideNavLink renderIcon={Fade} href="#">
+                    Link
+                  </SideNavLink>
+                  <SideNavLink renderIcon={Fade} href="#">
+                    Link
+                  </SideNavLink>
+                  <SideNavDivider />
+                  <SideNavLink renderIcon={Fade} href="#">
+                    Link
+                  </SideNavLink>
+                  <SideNavLink renderIcon={Fade} href="#">
+                    Link
+                  </SideNavLink>
+                </SideNavItems>
+              </SideNav>
+              <Theme theme="white">
+                <StoryContent />
+              </Theme>
+            </>
+          )}
+        />
+      </Theme>
+    </div>
+  );
+};
+
+Default.parameters = {
+  controls: { disable: true },
+  actions: { disable: true },
 };
 
 /**
- * Story for SideNav
+ * Story for UIShell with TreeView
  * @returns {React.ReactElement} The JSX for the story
  */
-export const Default = () => (
+export const DefaultWithTreeView = () => (
   <div>
     <Theme theme="g100">
       <HeaderContainer
@@ -176,16 +298,11 @@ export const Default = () => (
   </div>
 );
 
-Default.parameters = {
-  controls: { disable: true },
-  actions: { disable: true },
-};
-
 /**
  * Story for SideNav
  * @returns {React.ReactElement} The JSX for the story
  */
-export const SideNavFixed = () => (
+export const SideNavStory = () => (
   <SideNav
     isFixedNav
     expanded={true}
@@ -220,6 +337,37 @@ export const SideNavFixed = () => (
         <SideNavMenuItem href="#">Link</SideNavMenuItem>
         <SideNavMenuItem href="#">Link</SideNavMenuItem>
         <SideNavMenuItem href="#">Link</SideNavMenuItem>
+      </SideNavMenu>
+    </SideNavItems>
+  </SideNav>
+);
+
+/**
+ * Story for SideNav w/TreeView
+ * @returns {React.ReactElement} The JSX for the story
+ */
+export const ThirdMenuTest = () => (
+  <SideNav
+    isFixedNav
+    expanded={true}
+    isChildOfHeader={false}
+    aria-label="Side navigation">
+    <SideNavItems>
+      <SideNavLink renderIcon={Add} href="#">
+        Link level 1
+      </SideNavLink>
+      <SideNavMenu
+        defaultExpanded={true}
+        renderIcon={Fade}
+        title="Sub-menu level 1">
+        <SideNavMenuItem href="#">Link level 2</SideNavMenuItem>
+        <SideNavMenuItem href="#">Link level 2</SideNavMenuItem>
+        <SideNavMenu title="Sub-menu level 2" defaultExpanded={true}>
+          <SideNavMenuItem isActive href="#">
+            Link level 3
+          </SideNavMenuItem>
+          <SideNavMenuItem href="#">Link level 3</SideNavMenuItem>
+        </SideNavMenu>
       </SideNavMenu>
     </SideNavItems>
   </SideNav>
@@ -324,33 +472,53 @@ export const SideNavWithTreeView = () => (
   </div>
 );
 
+SideNavStory.storyName = 'SideNav';
+
 /**
- * Story for SideNav w/TreeView
+ * Story for HeaderPanel
+ * @param {object} args Storybook args that control component props
  * @returns {React.ReactElement} The JSX for the story
  */
-export const ThirdMenuTest = () => (
+export const HeaderPanelStory = (args) => <HeaderPanel {...args} />;
+HeaderPanelStory.storyName = 'HeaderPanel';
+
+HeaderPanelStory.args = {
+  expanded: true,
+};
+
+HeaderPanelStory.argTypes = {
+  expanded: { control: 'boolean' },
+};
+
+/**
+ * Story for SideNav
+ * @returns {React.ReactElement} The JSX for the story
+ */
+export const SideNavPanel = () => (
   <SideNav
-    isFixedNav
-    expanded={true}
+    navType={SIDE_NAV_TYPE.PANEL}
     isChildOfHeader={false}
     aria-label="Side navigation">
     <SideNavItems>
-      <SideNavLink renderIcon={Add} href="#">
-        Link level 1
+      <SideNavLink renderIcon={Fade} href="#">
+        Link
       </SideNavLink>
-      <SideNavMenu
-        defaultExpanded={true}
-        renderIcon={Fade}
-        title="Sub-menu level 1">
-        <SideNavMenuItem href="#">Link level 2</SideNavMenuItem>
-        <SideNavMenuItem href="#">Link level 2</SideNavMenuItem>
-        <SideNavMenu title="Sub-menu level 2" defaultExpanded={true}>
-          <SideNavMenuItem isActive href="#">
-            Link level 3
-          </SideNavMenuItem>
-          <SideNavMenuItem href="#">Link level 3</SideNavMenuItem>
-        </SideNavMenu>
-      </SideNavMenu>
+      <SideNavLink renderIcon={Fade} href="#">
+        Link
+      </SideNavLink>
+      <SideNavLink renderIcon={Fade} href="#">
+        Link
+      </SideNavLink>
+      <SideNavLink renderIcon={Fade} href="#">
+        Link
+      </SideNavLink>
+      <SideNavDivider />
+      <SideNavLink renderIcon={Fade} href="#">
+        Link
+      </SideNavLink>
+      <SideNavLink renderIcon={Fade} href="#">
+        Link
+      </SideNavLink>
     </SideNavItems>
   </SideNav>
 );